fix : 회원 탈퇴시 비밀번호 검증 단계 추가

See merge request cloud_track/class_03/web_project3/team02/chillvibe-backend!32

+++ src/main/java/com/chillvibe/chillvibe/domain/user/service/UserServiceImpl.java
@@ -5,6 +5,7 @@ import com.chillvibe.chillvibe.domain.hashtag.service.HashtagService;
 import com.chillvibe.chillvibe.domain.user.dto.JoinRequestDto;
 import com.chillvibe.chillvibe.domain.user.dto.PasswordUpdateRequestDto;
 import com.chillvibe.chillvibe.domain.user.dto.ReAuthResponseDto;
+import com.chillvibe.chillvibe.domain.user.dto.UserDeleteRequestDto;
 import com.chillvibe.chillvibe.domain.user.dto.UserInfoResponseDto;
 import com.chillvibe.chillvibe.domain.user.dto.UserUpdateRequestDto;
 import com.chillvibe.chillvibe.domain.user.entity.User;
@@ -189,13 +190,19 @@ public class UserServiceImpl implements UserService {
   }
 
   @Transactional
-  public void softDeleteUser() {
+  public void softDeleteUser(UserDeleteRequestDto userDeleteRequestDto) {
     Long userId = userUtil.getAuthenticatedUserId();
 
     User user = userRepository.findById(userId)
         .orElseThrow(() -> new ApiException(ErrorCode.USER_NOT_FOUND));
 
-    userRepository.delete(user);
+    String password = userDeleteRequestDto.getPassword();
+    if (bCryptPasswordEncoder.matches(password, user.getPassword())) {
+      userRepository.delete(user);
+    } else {
+      throw new ApiException(ErrorCode.INVALID_PASSWORD);
+    }
+
     performLogout(request, response);
   }
